requirements.txt
pytube~=11.0.1
pillow